Refactor job wrapper/runner utilities for reuse in extended set metadata.
