+++ lib/galaxy/jobs/__init__.py
@@ -16,7 +16,6 @@ import time
 import traceback
 from abc import ABCMeta, abstractmethod
 from json import loads
-from tempfile import NamedTemporaryFile
 from xml.etree import ElementTree
 
 import six
@@ -27,6 +26,7 @@ import galaxy
 from galaxy import model, util
 from galaxy.datatypes import sniff
 from galaxy.exceptions import ObjectInvalid, ObjectNotFound
+from galaxy.job_execution.output_collect import collect_extra_files
 from galaxy.jobs.actions.post import ActionBox
 from galaxy.jobs.mapper import JobMappingException, JobRunnerMapper
 from galaxy.jobs.runners import BaseJobRunner, JobState
@@ -1422,17 +1422,7 @@ class JobWrapper(HasResourceParameters):
             dataset.dataset.uuid = context['uuid']
         self.__update_output(job, dataset)
         if not purged:
-            self._collect_extra_files(dataset.dataset, self.working_directory)
-        # Handle composite datatypes of auto_primary_file type
-        if dataset.datatype.composite_type == 'auto_primary_file' and not dataset.has_data():
-            try:
-                with NamedTemporaryFile(mode='w') as temp_fh:
-                    temp_fh.write(dataset.datatype.generate_primary_file(dataset))
-                    temp_fh.flush()
-                    self.object_store.update_from_file(dataset.dataset, file_name=temp_fh.name, create=True)
-                    dataset.set_size()
-            except Exception as e:
-                log.warning('Unable to generate primary composite file automatically for %s: %s', dataset.dataset.id, e)
+            collect_extra_files(self.object_store, dataset, self.working_directory)
         if job.states.ERROR == final_job_state:
             dataset.blurb = "error"
             if not implicit_collection_jobs:
@@ -1705,31 +1695,6 @@ class JobWrapper(HasResourceParameters):
                 self.object_store.delete(self.get_job(), base_dir='job_work', entire_dir=True, dir_only=True, obj_dir=True)
         except Exception:
             log.exception("Unable to cleanup job %d", self.job_id)
-
-    def _collect_extra_files(self, dataset, job_working_directory):
-        object_store = self.app.object_store
-        store_by = getattr(object_store, "store_by", "id")
-        file_name = "dataset_%s_files" % getattr(dataset, store_by)
-        temp_file_path = os.path.join(job_working_directory, file_name)
-        extra_dir = None
-        try:
-            # This skips creation of directories - object store
-            # automatically creates them.  However, empty directories will
-            # not be created in the object store at all, which might be a
-            # problem.
-            for root, dirs, files in os.walk(temp_file_path):
-                extra_dir = root.replace(job_working_directory, '', 1).lstrip(os.path.sep)
-                for f in files:
-                    self.object_store.update_from_file(
-                        dataset,
-                        extra_dir=extra_dir,
-                        alt_name=f,
-                        file_name=os.path.join(root, f),
-                        create=True,
-                        preserve_symlinks=True
-                    )
-        except Exception as e:
-            log.debug("Error in collect_associated_files: %s" % (e))
 
     def _collect_metrics(self, has_metrics, job_metrics_directory=None):
         job = has_metrics.get_job()
